data_unifier/requirements.txt
python-magic-bin>=0.4.14  # Windows
pandas>=1.3.0
openpyxl>=3.0.0
camelot-py[base]>=0.10.0
pdfminer.six>=20200517
tabulate>=0.8.0